requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
langchain
langchain-community
langchain-google-genai
faiss-cpu==1.8.0
sentence-transformers==3.0.1
pypdf==4.2.0
python-dotenv==1.0.1
pydantic==2.7.4
httpx==0.27.0
